+++ src/core/history.py
@@ -31,14 +31,16 @@ class History(object):
         )
 
     def set_limit(self, limit: int):
-        self.limit: int = int(limit)
+        self.limit: int = limit
         if len(self.next) > limit:
             self.next = self.next[len(self.next)-limit:]
         if len(self.prev) > limit:
             self.prev = self.prev[len(self.prev)-limit:]
 
     def next_dir(self) -> Item:
-        self.stack_append(self.prev)
+        if len(self.prev) >= self.limit:
+            self.prev = self.prev[len(self.prev)-self.limit-1:]
+        self.prev.append(self.curr)
 
         self.curr = self.next.pop()
         ag.signals_.user_action_signal.emit(
@@ -47,18 +49,15 @@ class History(object):
         return self.curr
 
     def prev_dir(self) -> Item:
-        self.stack_append(self.next)
+        if len(self.next) >= self.limit:
+            self.next = self.next[len(self.next)-self.limit-1:]
+        self.next.append(self.curr)
 
         self.curr = self.prev.pop()
         ag.signals_.user_action_signal.emit(
             f'enable_next_prev/yes,{self.has_prev()}'
         )
         return self.curr
-
-    def stack_append(self, stack):
-        if len(stack) == self.limit:
-            stack = stack[:-(self.limit-1)]
-        stack.append(self.curr)
 
     def set_file_id(self, row_no: int):
         '''
@@ -76,7 +75,9 @@ class History(object):
 
     def add_item(self, path, file_id):
         if self.curr:
-            self.stack_append(self.prev)
+            if len(self.prev) >= self.limit:
+                self.prev = self.prev[len(self.prev)-self.limit-1:]
+            self.prev.append(self.curr)
         self.next.clear()
         ag.signals_.user_action_signal.emit(
             f'enable_next_prev/no,{self.has_prev()}'