+++ src/main.py
@@ -14,6 +14,14 @@ from PyQt6.QtWidgets import QApplication, QWidget
 from .core import utils, app_globals as ag
 from .core.sho import shoWindow
 
+if sys.platform.startswith("win"):
+    from .core import win_win as win_activate
+elif sys.platform.startswith("linux"):
+    from .core import linux_win as win_activate
+else:
+    raise ImportError(f"doesn't support {sys.platform} system")
+
+
 app: QApplication = None
 
 def app_name() -> str:
@@ -21,11 +29,9 @@ def app_name() -> str:
 
 def app_version() -> str:
     """
-    if it changes then also change in "pyproject.toml" file
-    what the shit is this versioning support API
-    it's easier to hardcode it everywhere
+    if version changed here then also change it in the "pyproject.toml" file
     """
-    return '0.9.3'
+    return '0.9.4'
 
 @pyqtSlot(QWidget, QWidget)
 def tab_pressed():
@@ -57,20 +63,15 @@ def main():
         lock_file = QLockFile(QDir.tempPath() + '/fileo.lock')
         # logger.info(f'{lock_file.fileName()}')
         if not lock_file.tryLock():
-            if lock_file.error() is QLockFile.LockError.LockFailedError:
-                res = lock_file.getLockInfo()
-                # logger.info(res)
-                import platform
-                if platform.system() == 'Windows':
-                    from pywinauto import Application
-                    running_app = Application().connect(process=res[1])
-                    running_app.top_window().set_focus()
-                elif platform.system() == 'Linux':
-                    pass
-                elif platform.system() == 'Darwin':
-                    pass
+            ag.single_instance = utils.get_app_setting("SINGLE_INSTANCE", False)
+            if ag.single_instance:
+                if lock_file.error() is QLockFile.LockError.LockFailedError:
+                    res = lock_file.getLockInfo()
+                    win_activate.activate(res)
 
-            sys.exit(0)
+                sys.exit(0)
+            else:
+                ag.db['restore'] = False
 
         global app
         app = QApplication([])